+++ app/src/main/java/com/example/mobileterm/iCalendarFragment.java
@@ -5,7 +5,7 @@ import static android.content.Context.MODE_NO_LOCALIZED_COLLATORS;
 import android.os.Bundle;
 
 import androidx.annotation.Nullable;
-import androidx.fragment.app.Fragment;
+
 
 import android.view.LayoutInflater;
 import android.view.View;
@@ -14,8 +14,6 @@ import android.annotation.SuppressLint;
 import android.content.Context;
 import android.graphics.Color;
 import android.graphics.drawable.Drawable;
-import android.os.Bundle;
-import android.view.View;
 import android.widget.Button;
 import android.widget.CalendarView;
 import android.widget.EditText;
@@ -26,6 +24,7 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.content.ContextCompat;
+import androidx.fragment.app.Fragment;
 
 import com.prolificinteractive.materialcalendarview.CalendarDay;
 import com.prolificinteractive.materialcalendarview.DayViewDecorator;
@@ -118,8 +117,6 @@ public class iCalendarFragment extends Fragment {
                 textView2.setVisibility(View.VISIBLE);
             }
         });
-
-
 
 
         return rootView;
